Split kotlin init & settings script tapi tests across supported versions
