+++ subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r48/KotlinSettingsScriptModelCrossVersionSpec.groovy
@@ -14,54 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.kotlin.dsl.tooling.builders.r41
+package org.gradle.kotlin.dsl.tooling.builders.r48
 
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
-import org.gradle.integtests.tooling.fixture.ToolingApiVersion
-
 import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
 
-import static org.junit.Assert.assertThat
 
-
-@ToolingApiVersion(">=4.1")
-@TargetGradleVersion(">=4.1")
+@TargetGradleVersion(">=4.8")
 class KotlinSettingsScriptModelCrossVersionSpec extends AbstractKotlinScriptModelCrossVersionTest {
-
-    def "can fetch classpath of settings script"() {
-
-        given:
-        withBuildSrc()
-
-        and:
-        def settingsDependency = withFile("settings-dependency.jar")
-        def settings = withSettings("""
-            buildscript {
-                dependencies {
-                    classpath(files("${normalizedPathOf(settingsDependency)}"))
-                }
-            }
-        """)
-
-        and:
-        def projectDependency = withFile("project-dependency.jar")
-        file("build.gradle") << """
-            buildscript {
-                dependencies {
-                    classpath(files("${normalizedPathOf(projectDependency)}"))
-                }
-            }
-        """
-
-        when:
-        def classPath = canonicalClassPathFor(projectDir, settings)
-
-        then:
-        assertContainsBuildSrc(classPath)
-        assertContainsGradleKotlinDslJars(classPath)
-        assertIncludes(classPath, settingsDependency)
-        assertExcludes(classPath, projectDependency)
-    }
 
     def "can fetch classpath of settings script plugin"() {
 
@@ -97,29 +57,5 @@ class KotlinSettingsScriptModelCrossVersionSpec extends AbstractKotlinScriptMode
         assertContainsGradleKotlinDslJars(classPath)
         assertIncludes(classPath, settingsDependency)
         assertExcludes(classPath, projectDependency)
-    }
-
-    def "sourcePath includes buildSrc source roots"() {
-
-        given:
-        withKotlinBuildSrc()
-        def settings = withSettings("""include(":sub")""")
-
-        expect:
-        assertThat(
-            sourcePathFor(settings),
-            matchesProjectsSourceRoots(withMainSourceSetJavaKotlinIn("buildSrc")))
-    }
-
-    def "sourcePath includes buildSrc project dependencies source roots"() {
-
-        given:
-        def sourceRoots = withMultiProjectKotlinBuildSrc()
-        def settings = withSettings("""include(":sub")""")
-
-        expect:
-        assertThat(
-            sourcePathFor(settings),
-            matchesProjectsSourceRoots(sourceRoots))
     }
 }
